requirements.txt
django
markdown 
python-docx 
xhtml2pdf 
beautifulsoup4